add test case for switch off ec

+++ test/triton_server.sh
@@ -7,7 +7,7 @@ notExistContainer="None"
 retryCount=3
 dockername=$1
 mode_repo=$2
-backend_repo=$3
+model_name=$3
 
 
 function GetContainerStatus(){
@@ -34,12 +34,22 @@ function StopContainer(){
 }
 
 function StartContainer(){
+if [ $model_name == 'dlrm' ]
+then
 nohup docker run --gpus=4 --rm  -p 8000:8000 -p 8001:8001 -p 8002:8002  \
 -v $mode_repo:/model $dockername \
 tritonserver --model-repository=/model/ --load-model=dlrm --load-model=dlrm_test --model-control-mode=explicit \
 --backend-directory=/usr/local/hugectr/backends/ \
---backend-config=hugectr,ps=/model/ps.json  \
---backend-config=hugectr,supportlonglong=true &
+--backend-config=hugectr,ps=/model/ps.json  &
+fi
+if [ $model_name == 'wdl' ]
+then
+nohup docker run --gpus=4 --rm  -p 8000:8000 -p 8001:8001 -p 8002:8002  \
+-v $mode_repo:/model $dockername \
+tritonserver --model-repository=/model/ --load-model=wdl --model-control-mode=explicit \
+--backend-directory=/usr/local/hugectr/backends/ \
+--backend-config=hugectr,ps=/model/ps_cpu.json  &
+fi
 
  echo "starting triton....."
  sleep 900
